+++ Marisa.Plugin.Shared/Osu/Entity/Score/Statistics.cs
@@ -6,20 +6,20 @@ namespace Marisa.Plugin.Shared.Osu.Entity.Score;
 public class Statistics
 {
     [JsonProperty("count_100", Required = Required.Always)]
-    public long Count100 { get; set; }
+    public int Count100 { get; set; }
 
     [JsonProperty("count_300", Required = Required.Always)]
-    public long Count300 { get; set; }
+    public int Count300 { get; set; }
 
     [JsonProperty("count_50", Required = Required.Always)]
-    public long Count50 { get; set; }
+    public int Count50 { get; set; }
 
     [JsonProperty("count_geki", Required = Required.Always)]
-    public long CountGeki { get; set; }
+    public int CountGeki { get; set; }
 
     [JsonProperty("count_katu", Required = Required.Always)]
-    public long CountKatu { get; set; }
+    public int CountKatu { get; set; }
 
     [JsonProperty("count_miss", Required = Required.Always)]
-    public long CountMiss { get; set; }
+    public int CountMiss { get; set; }
 }
